+++ plaid/__init__.py
@@ -8,7 +8,7 @@
 """
 
 
-__version__ = "9.1.0"
+__version__ = "9.1.1"
 
 # import ApiClient
 from plaid.api_client import ApiClient